+++ CHANGELOG.md
@@ -50,6 +50,8 @@ at locations that better optimize for object storage.
 - The Helm chart now supports Pod Disruption Budgets (PDBs) for Polaris components. This allows users to define
   the minimum number of pods that must be available during voluntary disruptions, such as node maintenance.
 
+- Feature configuration `PURGE_VIEW_METADATA_ON_DROP` was added to allow dropping views without purging their metadata files.
+
 ### Changes
 
 - Polaris Management API clients must be prepared to deal with new attributes in `AwsStorageConfigInfo` objects.